Added Manifest section
